+++ jeap-messaging-infrastructure/src/test/java/ch/admin/bit/jeap/messaging/kafka/log/TestEvent.java
@@ -66,11 +66,6 @@ public class TestEvent implements AvroMessage {
             public String getVersion() {
                 return "1.0.0";
             }
-
-            @Override
-            public String getVariant() {
-                return null;
-            }
         };
     }