+++ jeap-messaging-model/src/main/java/ch/admin/bit/jeap/messaging/model/MessageType.java
@@ -5,5 +5,7 @@ public interface MessageType {
 
     String getVersion();
 
-    String getVariant();
+    default String getVariant() {
+        return null;
+    }
 }